+++ src/index.html
@@ -15,7 +15,7 @@
   </head>
   <body>
     <!-- ❗️❗️❗️ Add partials like so for every page -->
-    <include src="./partials/example.html"></include>
+    <!-- <include src="./partials/example.html"></include> -->
 
     <include src="./partials/header/header.html"></include>
     <include src="./partials/home_page/home_page.html"></include>
