+++ src/partials/example.html
@@ -1,12 +1,17 @@
-<div class="section">
+<!-- <div class="section">
   <div class="container">
     <h1 class="title">Hello world!</h1>
   </div>
 
   <label class="switch">
     <p>Ligtn</p>
-    <input class="toggle-theme" type="checkbox" />
+
     <span class="slider round"></span>
     <p>Dark</p>
   </label>
 </div>
+
+
+<ul>
+  <li class="toggle-theme">toggle</li>
+</ul> -->
