+++ src/sass/_example.scss
@@ -1,4 +1,4 @@
-.dark .container {
-  background: black;
-  color: rgb(255, 255, 255);
-}
+// .dark .container {
+//   background: black;
+//   color: rgb(255, 255, 255);
+// }
